extra_tests: consolidate the _pickle snippets into one file
Merge the fourteen _pickle_* accelerator snippets into a single
_pickle.py. The sections share the dumps/loads/roundtrip helpers and run
sequentially; pickled classes stay at module scope so pickle-by-reference
still resolves. Sections needing the high-level pickle module (the module
flip, follow-ups, persistent_id) are gated on a HAVE_PICKLE flag,
replacing _pickle_module.py's standalone SystemExit skip, so an unpinned
stdlib skips them instead of aborting the whole file.
